# Conflicts:
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/.vs/Projet_Huchon_Salemi_3I/DesignTimeBuild/.dtbcache.v2
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/.vs/Projet_Huchon_Salemi_3I/v16/.suo
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/App.config
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/DAO/PersonneDAO.cs
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/MainWindow.xaml.cs
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/bin/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I.dll
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/bin/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I.dll.config
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/bin/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I.dll.locked
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/bin/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I.pdb
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/obj/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I.csproj.CoreCompileInputs.cache
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/obj/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I.dll
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/obj/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I.pdb
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/obj/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I_MarkupCompile.cache
#	Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/obj/Debug/netcoreapp3.1/Projet_Huchon_Salemi_3I_MarkupCompile.i.cache

+++ Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/Projet_Huchon_Salemi_3I/DAO/BaladeDAO.cs
@@ -27,7 +27,7 @@ namespace Projet_Huchon_Salemi_3I.DAO
             return false;
         }
 
-        public override Balade Find(int num)
+        public override Balade Find(decimal num)
         {
             Balade balade = null;
             try
